backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic==2.9.2
python-multipart==0.0.9
httpx==0.27.2
python-dotenv==1.0.1
openai>=1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
beautifulsoup4==4.12.3
